requirements.txt
# Core Flask and Extensions
Flask==3.0.3
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
Flask-Excel==0.0.7

# Database and Cache
SQLAlchemy==2.0.30
redis==5.0.1
greenlet==3.0.3

# Authentication and Security
bcrypt==4.0.1
passlib==1.7.4
PyJWT==2.8.0
email-validator==2.1.1

# Task Queue
celery==5.3.6
kombu==5.3.5

# Data Processing and Visualization
matplotlib==3.8.3
seaborn==0.13.2
pandas==2.2.1
numpy==1.26.4

# Monitoring and Logging
sentry-sdk[flask]==1.40.6
prometheus-flask-exporter==0.23.0
python-json-logger==2.0.7

# Utilities
python-dateutil==2.9.0
pytz==2024.1
six==1.16.0
aniso8601==9.0.1
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
<<<<<<< Updated upstream
WTForms==3.1.2
=======
WTForms==3.1.2

# Testing
pytest==8.0.2
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# Development Tools
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.2

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
>>>>>>> Stashed changes
